pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
authors = [
    {name = "SK-415", email = "2967923486@qq.com"},
]
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.8,<4.0"
dependencies = [
    "click>=8.1.3",
    "httpx>=0.24.0",
    "nonebot-adapter-onebot>=2.2.3",
    "nonebot-plugin-apscheduler>=0.2.0",
    "nonebot2[fastapi]>=2.0.0rc4",
    "playwright>=1.33.0",
    "pydantic>=1.10.7",
    "python-dotenv>=1.0.0",
    "tortoise-orm[asyncpg]>=0.19.3",
    "bilireq>=0.2.5",
    "packaging>=23.1",
    "nonebot-plugin-guild-patch>=0.2.3",
    "msvc-runtime>=14.34.31931; sys_platform == \"win32\"",
]
name = "haruka-bot"
version = "1.5.3.post1"
description = "Push dynamics and live informations from bilibili to QQ. Based on nonebot2."
readme = "README.md"
keywords = ["nonebot", "nonebot2", "qqbot", "bilibili", "bot"]

[project.urls]
homepage = "https://github.com/SK-415/HarukaBot"
repository = "https://github.com/SK-415/HarukaBot/tree/master/src/plugins/haruka_bot"
documentation = "https://github.com/SK-415/HarukaBot#readme"

[tool.pdm.dev-dependencies]
dev = [
    "nonebot-plugin-gocqhttp>=0.6.4",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["haruka_bot", "nonebot_plugin_gocqhttp"]
plugin_dirs = []
builtin_plugins = []

[project.scripts]
hb = "haruka_bot.__main__:main"

[tool.pdm.build]
# includes = ["haruka_bot/web/build/**/*"]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
